Sources are splitted to use separate unit tests
