+++ .gitignore
@@ -1,2 +1,3 @@
 .build
 .vscode
+*.txt.user
